Purus Pasta sync

+++ scripts/__pbot/PBotGobAPI.py
@@ -19,6 +19,7 @@ class PBotGobAPI(object):
         gob = self.__session.PBotGobAPI().getPlayer()
         return PBotGob(gob) if gob is not None else None
 
+
     ## Get closest gob matching the specific regex pattern
     # @param resname regex pattern
     # @return closest PBotGob
@@ -52,7 +53,7 @@ class PBotGobAPI(object):
     # @param bx x-coord of B point
     # @param by y-coord of B point
     # @return list of PBotGobs in the area
-    def gobs_in_area(self, ax: int, ay: int, bx: int, by: int) -> List[PBotGob]:
+    def gobs_in_area(self, ax: float, ay: float, bx: float, by: float) -> List[PBotGob]:
         return [PBotGob(x) for x in self.__session.PBotGobAPI().gobsInArea(ax, ay, bx, by)]
